sensor_anomaly_detection/requirements.txt
scikit-learn
pandas
numpy
matplotlib
seaborn
umap-learn
scipy
